+++ broker_memory.go
