+++ presence_memory.go
